+++ test/datalevin/server/connection_test.clj
@@ -1,7 +1,9 @@
 (ns datalevin.server.connection-test
   (:require
-   [clojure.test :refer [deftest is use-fixtures]]
+   [clojure.test :refer [deftest is testing use-fixtures]]
+   [datalevin.constants :as c]
    [datalevin.core :as d]
+   [datalevin.remote :as remote]
    [datalevin.server :as server]
    [datalevin.test.core :refer [allocate-port db-fixture]]
    [datalevin.util :as u])
@@ -80,3 +82,71 @@
             (is (thrown? clojure.lang.ExceptionInfo (d/commit-kv-transaction tx)))
             (is (nil? (d/get-value kv "data" :key))))
           (finally (d/close-kv kv)))))))
+
+(deftest datalog-transaction-exposes-kv-operations-test
+  (with-server
+    (fn [_ uri]
+      (doseq [wal? [false true]]
+        (let [conn (d/create-conn (str uri "-mixed-" wal?)
+                                  {:value {:db/valueType :db.type/keyword}}
+                                  {:client-opts {:pool-size 3 :time-out 5000}
+                                   :kv-opts {:wal? wal?}})
+              kv (d/datalog-kv conn)]
+          (try
+            (is (= (d/dir kv) (d/dir (d/datalog-kv @conn))))
+            (d/open-dbi kv "state")
+            (d/open-list-dbi kv "items")
+            (doseq [outcome [:commit :abort :exception]]
+              (testing (str "WAL " wal? ", " outcome)
+                (d/transact! conn [{:db/id 1 :value :before}])
+                (d/transact-kv kv "state" [[:put :key :before]])
+                (d/clear-dbi kv "items")
+                (let [run (fn []
+                            (d/with-transaction [tx conn]
+                              (let [tx-kv (d/datalog-kv tx)]
+                                (d/transact! tx [[:db/add 1 :value :inside]])
+                                (d/transact-kv tx-kv "state" [[:put :key :inside]])
+                                (is (= :inside (d/get-value tx-kv "state" :key)))
+                                (is (= :inside (d/get-value (d/datalog-kv @tx)
+                                                           "state" :key)))
+                                (is (= [[:key :inside]]
+                                       (d/get-range tx-kv "state" [:all])))
+                                (is (= [:inside] (remote/get-values tx-kv "state" [:key])))
+                                (is (= :inside ((d/prepare-get-value tx-kv "state")
+                                                :key)))
+                                (d/put-list-items tx-kv "items" :key [1 2] :data :long)
+                                (is (= [1 2] (d/get-list tx-kv "items" :key :data :long)))
+                                (d/with-transaction [nested tx]
+                                  (d/with-transaction-kv [nested-kv (d/datalog-kv nested)]
+                                    (is (= :inside (d/get-value nested-kv "state" :key)))))
+                                (is (= :inside (:value (d/pull @tx [:value] 1))))
+                                (is (= :before (d/get-value kv "state" :key)))
+                                (is (= :before (:value (d/pull @conn [:value] 1))))
+                                (is (empty? (d/get-list kv "items" :key :data :long)))
+                                (case outcome
+                                  :abort (d/abort-transact tx)
+                                  :exception (throw (ex-info "rollback mixed writes" {}))
+                                  nil))))]
+                  (if (= outcome :exception)
+                    (is (thrown-with-msg? clojure.lang.ExceptionInfo
+                                          #"rollback mixed writes" (run)))
+                    (run)))
+                (let [expected (if (= outcome :commit) :inside :before)]
+                  (is (= expected (d/get-value kv "state" :key)))
+                  (is (= expected (:value (d/pull @conn [:value] 1)))))
+                (is (= (if (= outcome :commit) [1 2] [])
+                       (vec (d/get-list kv "items" :key :data :long))))))
+            ;; Copy-in writes and copy-out reads must use the Datalog runner too.
+            (d/open-dbi kv "bulk")
+            (let [n (inc c/+wire-datom-batch-size+)
+                  rows (mapv (fn [i] [i i]) (range n))]
+              (d/with-transaction [tx conn]
+                (let [tx-kv (d/datalog-kv tx)]
+                  (d/transact-kv tx-kv "bulk"
+                                 (mapv (fn [[k v]] [:put k v]) rows) :long :long)
+                  (is (= rows (d/get-range tx-kv "bulk" [:all] :long :long)))
+                  (is (zero? (d/entries kv "bulk")))))
+              (is (= n (d/entries kv "bulk"))))
+            (d/close conn)
+            (is (d/closed-kv? kv))
+            (finally (d/close conn))))))))